# Optimizing casting GenericView to primitives.