Summary:
When casting a genericView, the underlying reader is first ensured to have been created. 
while its only created once, the check that is created happens over and over for each row. 

When the generic view represents a primitive type, there is only one valid destination type, 
in that case we can ensure the casted reader when the vector reader is created and avoid the check 
for each row. The diff does this optimization.

after
```
============================================================================
[...]hmarks/ExpressionBenchmarkBuilder.cpp     relative  time/iter   iters/s
============================================================================
array_sum##generic_input_int                               55.21ms     18.11
array_sum##typed_int                                       53.03ms     18.86
array_sum##generic_input_double                            59.23ms     16.88
array_sum##typed_double                                    58.89ms     16.98
```

before:
```
============================================================================
[...]hmarks/ExpressionBenchmarkBuilder.cpp     relative  time/iter   iters/s
============================================================================
array_sum##generic_input_int                               73.16ms     13.67
array_sum##typed_int                                       55.23ms     18.10
array_sum##generic_input_double                            72.83ms     13.73
array_sum##typed_double                                    59.04ms     16.94
```

Reviewed By: Yuhta

Differential Revision: D52223370